Create f5_ltm_stats.py
Makes two REST API calls using the imported 'f5api_call.py' module, first call extracts the LTM pool stats and the second extracts the virtual server details. It then cross references the response output from both to ascertain which virtual servers are in use and which are inactive. It then allows you to write either or both to a .csv file.